+++ packages/amplify-cli/src/extensions/amplify-helpers/push-resources.ts
@@ -10,8 +10,17 @@ export async function pushResources(
   context: $TSContext,
   category?: string,
   resourceName?: string,
-  filteredResources?: { category: string, resourceName: string }[],
+  filteredResources?: { category: string; resourceName: string }[],
 ) {
+  if (context.parameters.options['iterative-rollback']) {
+    // validate --iterative-rollback with --force
+    if (context.parameters.options.force) {
+      throw new Error(
+        "'--iterative-rollback' and '--force' cannot be used together. Consider runnning 'amplify push --force' to iteratively rollback and redeploy.",
+      );
+    }
+    context.exeInfo.iterativeRollback = true;
+  }
   if (context.parameters.options.env) {
     const envName: string = context.parameters.options.env;
     const allEnvs = context.amplify.getAllEnvs();
@@ -25,13 +34,10 @@ export async function pushResources(
       });
 
       context.exeInfo.localEnvInfo = getEnvInfo();
-
       if (context.exeInfo.localEnvInfo.envName !== envName) {
         context.exeInfo.localEnvInfo.envName = envName;
-
         stateManager.setLocalEnvInfo(context.exeInfo.localEnvInfo.projectPath, context.exeInfo.localEnvInfo);
       }
-
       await initializeEnv(context);
     } else {
       const errMessage = "Environment doesn't exist. Please use 'amplify init' to create a new environment";
@@ -55,6 +61,9 @@ export async function pushResources(
   let continueToPush = context.exeInfo && context.exeInfo.inputParams && context.exeInfo.inputParams.yes;
 
   if (!continueToPush) {
+    if (context.exeInfo.iterativeRollback) {
+      context.print.info('The CLI will rollback the last known iterative deployment.');
+    }
     continueToPush = await context.amplify.confirmPrompt('Are you sure you want to continue?');
   }