+++ packages/amplify-e2e-core/src/init/amplifyPush.ts
@@ -1,4 +1,4 @@
-import { getCLIPath, KEY_DOWN_ARROW, nspawn as spawn } from '..';
+import { getCLIPath, KEY_DOWN_ARROW, CONTROL_C, nspawn as spawn } from '..';
 
 const pushTimeoutMS = 1000 * 60 * 20; // 20 minutes;
 
@@ -31,6 +31,37 @@ export function amplifyPushForce(cwd: string, testingWithLatestCodebase: boolean
           resolve();
         } else {
           reject(err);
+        }
+      });
+  });
+}
+
+/**
+ *
+ * @param cwd
+ * @param stackIdx
+ * @param testingWithLatestCodebase
+ * @returns
+ * Used to stop an iterative deployment
+ * Waits on the table stack to be complete and for the next stack to update in order to cancel the push
+ */
+export function cancelIterativeAmplifyPush(
+  cwd: string,
+  idx: { current: number; max: number },
+  testingWithLatestCodebase: boolean = false,
+): Promise<void> {
+  return new Promise((resolve, reject) => {
+    spawn(getCLIPath(testingWithLatestCodebase), ['push'], { cwd, stripColors: true, noOutputTimeout: pushTimeoutMS })
+      .wait('Are you sure you want to continue?')
+      .sendConfirmYes()
+      .wait(`Deploying (${idx.current} of ${idx.max})`)
+      .wait(/.*UPDATE_IN_PROGRESS GraphQLSchema*/)
+      .sendCtrlC()
+      .run((err: Error) => {
+        if (err && !/Killed the process as no output receive for/.test(err.message)) {
+          reject(err);
+        } else {
+          resolve();
         }
       });
   });
@@ -110,5 +141,20 @@ export function amplifyPushLayer(
         reject(err);
       }
     });
+  });
+}
+
+export function amplifyPushIterativeRollback(cwd: string, testingWithLatestCodebase: boolean = false) {
+  return new Promise((resolve, reject) => {
+    spawn(getCLIPath(testingWithLatestCodebase), ['push', '--iterative-rollback'], { cwd, stripColors: true })
+      .wait('Are you sure you want to continue?')
+      .sendLine('y')
+      .run((err: Error) => {
+        if (!err) {
+          resolve({});
+        } else {
+          reject(err);
+        }
+      });
   });
 }