+++ packages/amplify-provider-awscloudformation/src/utils/env-level-constructs.ts
@@ -5,6 +5,7 @@ import constants from '../constants';
 import { NetworkStack } from '../network/stack';
 import { getEnvironmentNetworkInfo } from '../network/environment-info';
 import { prePushCfnTemplateModifier } from '../pre-push-cfn-processor/pre-push-cfn-modifier';
+import { consolidateApiGatewayPolicies } from './consolidate-apigw-policies';
 
 const { ProviderName: providerName } = constants;
 
@@ -15,7 +16,11 @@ export async function createEnvLevelConstructs(context) {
 
   const updatedMeta = {};
 
-  Object.assign(updatedMeta, await createNetworkResources(context, stackName, hasContainers));
+  Object.assign(
+    updatedMeta,
+    await createNetworkResources(context, stackName, hasContainers),
+    consolidateApiGatewayPolicies(context, stackName),
+  );
 
   context.amplify.updateProvideramplifyMeta(providerName, updatedMeta);