+++ packages/amplify-dynamodb-simulator/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@1.18.2...amplify-dynamodb-simulator@1.18.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-dynamodb-simulator
+
+
+
+
+
+## [1.18.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@1.18.1...amplify-dynamodb-simulator@1.18.2) (2021-04-14)
+
+**Note:** Version bump only for package amplify-dynamodb-simulator
+
+
+
+
+
 ## [1.18.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@1.17.13...amplify-dynamodb-simulator@1.18.1) (2021-04-09)
 
 **Note:** Version bump only for package amplify-dynamodb-simulator