+++ packages/amplify-util-import/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.4.2...amplify-util-import@1.4.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-util-import
+
+
+
+
+
+## [1.4.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.4.1...amplify-util-import@1.4.2) (2021-04-14)
+
+**Note:** Version bump only for package amplify-util-import
+
+
+
+
+
 ## [1.4.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.3.15...amplify-util-import@1.4.1) (2021-04-09)
 
 **Note:** Version bump only for package amplify-util-import