+++ packages/graphql-elasticsearch-transformer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.11.3](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@4.11.2...graphql-elasticsearch-transformer@4.11.3) (2021-04-19)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
+## [4.11.2](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@4.11.1...graphql-elasticsearch-transformer@4.11.2) (2021-04-14)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
 ## [4.11.1](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@4.10.8...graphql-elasticsearch-transformer@4.11.1) (2021-04-09)
 
 **Note:** Version bump only for package graphql-elasticsearch-transformer